refactor renterd, do not show window on startup if configured
