+++ renterd/electron-src/state.ts
@@ -1,7 +1,24 @@
 import { ChildProcess } from 'child_process'
+import { BrowserWindow, Tray } from 'electron'
 
 export let state: {
-  process: ChildProcess | null
+  mainWindow: BrowserWindow | null
+  tray: Tray | null
+  daemon: ChildProcess | null
+  isQuitting: boolean
 } = {
-  process: null,
+  mainWindow: null,
+  tray: null,
+  isQuitting: false,
+  daemon: null,
+}
+
+export const system: {
+  isDarwin: boolean
+  isLinux: boolean
+  isWindows: boolean
+} = {
+  isDarwin: process.platform === 'darwin',
+  isLinux: process.platform === 'linux',
+  isWindows: process.platform === 'win32',
 }
